Migrate to next.js

+++ src/components/PhotoGrid.js
@@ -5,6 +5,7 @@
  * 
  * @author Robert Norwood
  * @date October, 2023 
+ * @last_modified May 16, 2025
  */
 
 import React, { useState, useEffect, useRef, useCallback } from 'react';
@@ -12,23 +13,24 @@ import {
     MDBBtn,
     MDBIcon,
 } from 'mdb-react-ui-kit';
+
 import { } from '@aws-amplify/ui-react';
 import { remove } from 'aws-amplify/storage';
 import { generateClient } from 'aws-amplify/api';
-import projectConfig from '../helpers/Config';
+import { IMAGEDELIVERYHOST } from '../helpers/Config';
 // Database
 import { imagesByAlbumsID } from '../graphql/queries';
 import { deleteImages as deleteImageMutation } from '../graphql/mutations';
+import ResponsiveGrid from "./ResponsiveGrid";
 
 // Components
 import { Lightbox } from "yet-another-react-lightbox";
 import Download from "yet-another-react-lightbox/plugins/download";
-import ResponsiveGrid from "./ResponsiveGrid";
 import Zoom from "yet-another-react-lightbox/plugins/zoom";
 import "yet-another-react-lightbox/styles.css";
-import Image from "./Image";
-// import "animate.css/animate.min.css";
-// import { AnimationOnScroll } from 'react-animation-on-scroll';
+import ImageWrapper from "./Image";
+
+import { breakpoints } from "./Home";
 
 // Inputs:
 // setFeaturedImg - callback to set an image as the albums featured image
@@ -42,127 +44,106 @@ const userGroupClient = generateClient({
     authMode: 'userPool'
 });
 
-/**
+
+// Number of images to load per fetch
+const numImagesToLoad = 10;
+
+
+
+
+/** 
  * @param {
  *  setFeaturedImg - callback for setting the feature image
  *  selectedAlbum - album object to display
- *  editMode - should user be able to set featured img or delete etc
+ *  editMode - should user be able to set featured img or delete etc. 
  *  signedIn - is there a user signed in
  * } 
  * @returns React Component
  */
 export default function PhotoGrid({ setFeaturedImg, selectedAlbum, editMode = false, signedIn = false }) {
 
-    const [open, setOpen] = React.useState(false);
 
-    // Tracks index for Lightbox
-    const [index, setIndex] = React.useState(0);
+
+    // Ref for loading state, avoids race condition
+    const isLoadingFetching = useRef(false);
+
+
+    // State for lightbox
+    const [open, setOpen] = useState(false);
+
+    // Tracks index for Lightbox    
+    const [index, setIndex] = useState(0);
 
     // Stores items to display in grid/lightbox
     const [items, setItems] = useState([]);
 
-    // Observer for infinite scroll
-    const observerTarget = useRef(null);
-
     // Holds next token for data
-    const [nextToken, setNextToken] = useState(null);
-    const [isLoading, setIsLoading] = useState(true);
+    const [nextToken, setNextToken] = useState(-1);
+    const [isLoading, setIsLoading] = useState(false);
 
     // Fetches next set of items when bottom of scroll is reached
     const fetchData = useCallback(async () => {
-        // async function fetchData(){
-        if (isLoading || !nextToken) return;
+
+        // Don't fetch more data if already loading or no more data to fetch
+        if (isLoadingFetching.current || nextToken === null) return;
+
 
         setIsLoading(true);
+        isLoadingFetching.current = true;
 
-        const res = await client.graphql({
-            query: imagesByAlbumsID,
-            variables: {
+        try {
+            const queryData = {
                 albumsID: selectedAlbum.id,
-                limit: 10,
-                nextToken: nextToken
-            },
-        });
+                limit: numImagesToLoad
+            }
+            if (nextToken !== -1) {
+                queryData.nextToken = nextToken;
+            }
 
-        setNextToken(res.data.imagesByAlbumsID.nextToken);
-        const new_items = [...items, ...res.data.imagesByAlbumsID.items].map((img, i) => {
-            img.index = i;
-            return img
-        });
-        setItems(new_items);
-        setIsLoading(false);
-        // }
-    }, [nextToken]);
+            let res = await client.graphql({
+                query: imagesByAlbumsID,
+                variables: queryData
+            });
 
-    // Fetches initial data upon load
+            if (!res || !res.data) {
+                console.error("Error fetching data", res);
+                setIsLoading(false);
+                return;
+            }
+
+            const newItems = res.data.imagesByAlbumsID.items.map((img, i) => {
+                img.index = i;
+                return img
+            });
+            setItems(items => [...items, ...newItems]);
+
+            setNextToken(res.data.imagesByAlbumsID.nextToken);
+        } catch (error) {
+            console.error("Error fetching data", error);
+        } finally {
+            setIsLoading(false);
+            isLoadingFetching.current = false;
+            return;
+
+        }
+
+    }, [nextToken, selectedAlbum.id]);
+
+
+    // Effect to trigger initial data fetch for an album
     useEffect(() => {
-        if (index < items.length - 1) {
+        if (nextToken === -1 && selectedAlbum.id && items.length === 0 && !isLoadingFetching.current) {
             fetchData();
         }
-    }, [index]);
-
-
-    // Initalizes intersection observer to call each time observer enters view
-    // useEffect(() => {
-    //     const observer = new IntersectionObserver(
-    //         entries => {
-    //             if (entries[0].isIntersecting) {
-    //                 fetchData();
-    //             }
-    //         },
-    //         // { threshold: 1 }
-    //     );
-
-    //     const obsCurrent = observerTarget.current;
-
-    //     if (obsCurrent && !isLoading) {
-    //         observer.observe(obsCurrent);
-    //     }
-
-    //     return () => {
-    //         if (obsCurrent) {
-    //             observer.unobserve(obsCurrent);
-    //         }
-    //     };
-    // }, [isLoading, fetchData]);
-
-    // https://dev.to/vishnusatheesh/exploring-infinite-scroll-techniques-in-react-1bn0
-
-    // 
-    useEffect(() => {
-        getImages();
-    }, []);
-
-    //     /*Breakpoints. Breakpoint will be set to the last value before window width. Index will be the number of columns
-    //   Example  breakpoints = [0 ,  350, 750, 900, 1300]
-    //         number columns = [0 ,   1 ,  2 , 3  ,   4 ]
-    //         Window with of 850 will have 2 columns. 2000 will have 4
+    }, [nextToken, selectedAlbum.id, items.length, fetchData]);
 
 
 
-    // Requests the first 10 images
-    async function getImages() {
-        setIsLoading(true);
-        // Pulls the image objects associated with the selected album
-        const res = await client.graphql({
-            query: imagesByAlbumsID,
-            variables: {
-                albumsID: selectedAlbum.id,
-                limit: 10
-            },
-        });
-        const imgs = res.data.imagesByAlbumsID.items.map((img, i) => {
-            img.index = i;
-            return img
-        });
-        const nextT = res.data.imagesByAlbumsID.nextToken;
-        setNextToken(nextT);
 
-        setItems(imgs);
-        setIsLoading(false);
-    }
+
 
     // Deletes image object and source image on AWS
+    // TODO(bobby): extract to a Model function
     async function deleteImage(image) {
         const newImages = items.filter((img) => img.id !== image.id);
         await remove({
@@ -177,40 +158,34 @@ export default function PhotoGrid({ setFeaturedImg, selectedAlbum, editMode = fa
     }
 
     // Ensures grid does not render if no items are in props
-    if (items.length === 0) return;
-
-    // Ensures each image has a unique index to map to the lightbox
-    // const items_with_index = items.map((item,i) => {
-    //   item.index = i;
-    //   return item;});
+    if (items.length === 0 && nextToken !== null) {
+        return;
+    } else if (items.length === 0 && nextToken === null) {
+        return (<div className='text-center'>No images found for this album</div>)
+    }
 
     // Slides object for lightbox doesn't hold full image object, just the url
     const slides = items.map((image) => {
         const urlNoSpaces = `${image.id}-${image.filename}`.replaceAll(' ', '%20');
         return ({
-            src: `https://${projectConfig.getValue('imageDeliveryHost')}/public/${image.id}-${image.filename}`,
+            src: `https://${IMAGEDELIVERYHOST}/public/${image.id}-${image.filename}`,
             alt: image.filename,
 
-            downloadUrl: `https://${projectConfig.getValue('imageDeliveryHost')}/public/${urlNoSpaces}`,
+            downloadUrl: `https://${IMAGEDELIVERYHOST}/public/${urlNoSpaces}`,
             width: image.width,
             height: image.height,
         });
     }
     );
 
-    // srcSet:[ 
-    //   { src: `https://${IMAGEDELIVERYHOST}/public/${urlNoSpaces}?width=1920`, width: 1920},
-    //   ],
-
-
-
 
     function confirmDeleteImage(image) {
-        if (window.confirm("Are you sure you want to delete this image?")) {
+        if (typeof window !== 'undefined' && window.confirm("Are you sure you want to delete this image?")) {
             deleteImage(image);
         }
     }
 
+    // Delete image button that renders when edit mode is on
     function DeleteImageWrapper(image) {
         if (!deleteImage || !editMode) {
             return;
@@ -242,9 +217,10 @@ export default function PhotoGrid({ setFeaturedImg, selectedAlbum, editMode = fa
     const responsiveProps = items.map((image, i) => (
         <div className='m-0 p-1' key={i}>
             <div className='bg-image hover-overlay position-relative'>
-                <Image
+                <ImageWrapper
                     img_obj={image}
                     className='img-fluid shadow-4'
+                    alt={image.filename}
                 />
                 <a type="button" >
                     <div className='mask overlay' onClick={() => (setOpen(true), setIndex(image.index))}
@@ -264,12 +240,11 @@ export default function PhotoGrid({ setFeaturedImg, selectedAlbum, editMode = fa
         <div className="d-flex photo-album">
             <ResponsiveGrid
                 items={responsiveProps}
-                breakpoints={[0, 350, 750, 1200]}
+                breakpoints={breakpoints}
                 loadNextItems={fetchData}
                 isLoading={isLoading}
                 setIsLoading={setIsLoading}
             />
-            {/* <div ref={observerTarget}></div> */}
             <Lightbox
                 index={index}
                 slides={slides}
